# docs/architecture/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "本机管理系统当前 HTTP 契约",
    "version": "2026-08-08-working-tree",
    "description": "从当前Flask路由提取。x-legacy-exception表示请求/响应尚未统一；字段继续以源码与测试为准。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "默认本机Flask"
    }
  ],
...
  "paths": {
...
    "/api/status": {
      "get": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_get_status_get",
        "summary": "GET /api/status",
        "x-source": "gateway/app.py:7069",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
    "/api/content/videos/sync": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_sync_content_videos_route_post",
        "summary": "POST /api/content/videos/sync",
        "x-source": "gateway/app.py:8241",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/content/videos": {
      "get": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_content_videos_route_get",
        "summary": "GET /api/content/videos",
        "x-source": "gateway/app.py:8249",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/content/brands": {
      "get": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_content_brands_route_get",
        "summary": "GET /api/content/brands",
        "x-source": "gateway/app.py:8253",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_create_content_brand_route_post",
        "summary": "POST /api/content/brands",
        "x-source": "gateway/app.py:8257",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/content/brands/{brand_id}": {
      "patch": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_rename_content_brand_route_patch",
        "summary": "PATCH /api/content/brands/{brand_id}",
        "x-source": "gateway/app.py:8266",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/content/copy": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_add_content_copy_route_post",
        "summary": "POST /api/content/copy",
        "x-source": "gateway/app.py:8287",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/content/copy/import": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_import_content_copy_route_post",
        "summary": "POST /api/content/copy/import",
        "x-source": "gateway/app.py:8303",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/content/brands/{brand_id}/copy": {
      "get": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_content_copy_route_get",
        "summary": "GET /api/content/brands/{brand_id}/copy",
        "x-source": "gateway/app.py:8322",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/publish/queue/manual-test": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_manual_publish_test_route_post",
        "summary": "POST /api/publish/queue/manual-test",
        "x-source": "gateway/app.py:8327",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/publish/queue/batch": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_batch_publish_queue_route_post",
        "summary": "POST /api/publish/queue/batch",
        "x-source": "gateway/app.py:8347",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/publish/queue/batches": {
      "get": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_batch_publish_runs_route_get",
        "summary": "GET /api/publish/queue/batches",
        "x-source": "gateway/app.py:8407",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/publish/queue/process-one": {
      "post": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_process_one_publish_task_route_post",
        "summary": "POST /api/publish/queue/process-one",
        "x-source": "gateway/app.py:8412",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/publish/results": {
      "get": {
        "tags": [
          "Legacy Gateway"
        ],
        "operationId": "legacy_gateway_publish_results_route_get",
        "summary": "GET /api/publish/results",
        "x-source": "gateway/app.py:8440",
        "x-legacy-exception": true,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "认证模式Flask Session；local-direct模式由loopback/Host guard保护。"
      }
    },
...
    "responses": {
      "GenericSuccess": {
        "description": "当前成功响应；Legacy路由形状可能不统一。",
        "content": {
          "application/json": {
            "schema": {}
          }
        }
      },
      "BadRequest": {
        "description": "请求无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未认证",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "权限或本机边界拒绝",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "revision或状态冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Validation": {
        "description": "业务校验失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Internal": {
        "description": "内部错误，消息固定且脱敏",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}